Trim trace provider template helpers (#30)
Move replay template extraction and substitution helpers out of
`trace_provider.rs` into a sibling support module so the provider stays
under the repository file-size cap.

Update the replay cursor docs to describe post-exhaustion advancement
accurately, and use an `rstest` temporary-directory fixture in cleanup
support tests to remove repeated setup code.
