+++ tests/support/e2e.rs
@@ -11,6 +11,7 @@ pub mod routines;
 pub mod test_channel;
 pub mod test_rig;
 pub mod trace_provider;
+mod trace_template_utils;
 pub mod trace_types;
 mod trace_types_patch;
 mod trace_types_runtime;
